#1574 wip - icon build stuff to a separate file
