+++ .editorconfig
@@ -3,7 +3,7 @@ root = true
 [*.{js,json,purs,less}]
 charset = utf-8
 indent_style = space
-ident_size = 2
+indent_size = 2
 end_of_line = lf
 trim_trailing_whitespace = true
 insert_final_newline = true
