+++ gulpfile.js
@@ -2,23 +2,20 @@
 
 var gulp = require("gulp"),
     header = require("gulp-header"),
-    cache = require("gulp-cached"),
-    cheerio = require("gulp-cheerio"),
     contentFilter = require("gulp-content-filter"),
-    inject = require("gulp-inject"),
     purescript = require("gulp-purescript"),
-    svgSprite = require("gulp-svg-sprite"),
     webpack = require("webpack-stream"),
     rimraf = require("rimraf"),
     fs = require("fs"),
-    changeCase = require("change-case"),
-    glob = require("glob"),
     trimlines = require("gulp-trimlines"),
     less = require("gulp-less"),
     sequence = require("run-sequence"),
     replace = require("gulp-replace"),
     footer = require("gulp-footer"),
     path = require("path");
+
+const { injectIconsIntoHTML, createIconPureScript } = require("./script/icons")
+
 
 var slamDataSources = [
   "src/**/*.purs",
@@ -244,155 +241,8 @@ gulp.task("less", function() {
 mkWatch("watch-less", "less", ["less/**/*.less"]);
 
 
-gulp.task("svg-icons", () => {
-  // TODO: figure out how to use this
-  const iconAttribution = {};
-
-  const pathRemoveAttrs = [
-    "color",
-    "font-family",
-    "font-weight",
-    "overflow",
-    "style"
-  ];
-
-  const svgSymbols =
-    gulp.src("icons/**/*.svg")
-      .pipe(cache("iconing"))
-      // adjust SVGs from Project Noun
-      .pipe(cheerio({
-        parserOptions: { xmlMode: true },
-        run: ($, file) => {
-          const name = changeCase.titleCase(path.parse(file.path).name);
-          const svg = $("svg");
-          // inject the title for accessibility
-          svg.prepend(`<title>${name}</title>`);
-          const text = svg.find("text");
-          // not the best check, but I'm assuming these are Project Noun
-          // files
-          if (text.length > 0) {
-            const v = svg.attr("viewBox").match(/\d+/g);
-            if (v.length === 4) {
-              // squaring up the viewbox based on smaller value
-              const s = Math.min(v[2], v[3]);
-              svg.attr("viewBox", `0 0 ${s} ${s}`);
-            }
-            // push icon into the attribution object
-            const a = text.map(function() { return $(this).text() }).get().join(" ");
-            iconAttribution[a] = a in iconAttribution
-              ? iconAttribution[a].concat(name)
-              : iconAttribution[a] = [name];
-          }
-          text.remove();
-          // these irrelevant attrs need to be removed
-          const path_ = svg.find("path");
-          pathRemoveAttrs.forEach((attr) => path_.removeAttr(attr));
-        }
-      }))
-      // Transform the SVGs into a symbol sprite
-      .pipe(svgSprite({
-        svg: {
-          xmlDeclaration: false,
-          doctypeDeclaration: false,
-          rootAttributes: {
-            "aria-hidden": "true",
-            style: "display:none"
-          }
-        },
-        shape: {
-          dimensions: { maxWidth: 100, maxHeight: 100, precision: 3 },
-          id: {
-            // `file.stem` not supported in this vinyl version :/
-            // using this to override the default behavior of adding
-            // the directory to the id
-            generator: (name, file) => `sd-icon--${path.parse(name).name.replace(/\s/g, "-")}`
-          }
-        },
-        mode: {
-          symbol: {
-            render: { css: false, scss: false },
-          }
-        }
-      }))
-
-
-  // inject symbols into html files
-  gulp.src("html/**/*.html")
-    .pipe(inject(svgSymbols, {
-      starttag: "<!-- icon-symbols -->",
-      endtag: "<!-- /icon-symbols -->",
-      removeTags: true,
-      transform: (filePath, file) => file.contents.toString("utf8")
-    }))
-    .pipe(gulp.dest('./public'));
-});
-
-
-gulp.task("icon-purs", () =>
-  glob("icons/**/*.svg", {}, (err, files) => {
-    if (err) {
-      throw err;
-      return;
-    }
-
-    const { exports, html } = files.reduce((acc, file) => {
-      // get unique file names
-      const p = path.parse(file);
-      if (p == null) {
-        console.error("[icon-purs-adt]", file, "didn't parse.");
-}
-      const name = p.name.replace(/\s/g, "-");
-      if (!acc.includes(name)) {
-        acc.push(name);
-      } else {
-        console.warn("[icon-purs-adt]:", name, "from", file, "is a duplicate name & is being ignored.");
-      }
-      return acc;
-    }, []).reduce((acc, name) => {
-      // make an array for exports and one for the HTML
-      const camelName = changeCase.camelCase(name);
-      const html = `${camelName} ∷ ∀ p i. H.HTML p i
-${camelName} = iconHelper "${name}"`;
-      acc.exports.push(camelName);
-      acc.html.push(html);
-      return acc;
-    }, { exports: [], html: [] });
-
-    const pursData = `module SlamData.Render.Icon
-  ( ${exports.join("\n  , ")}
-  ) where
-
-import SlamData.Prelude
-
-import Halogen.HTML as H
-import Halogen.HTML.Properties as P
-
-iconHelper ∷ ∀ p i. String → H.HTML p i
-iconHelper s =
-  let
-    svgElem = H.elementNS $ H.Namespace "http://www.w3.org/2000/svg"
-    xlinkAttr = H.attrNS $ H.Namespace "http://www.w3.org/1999/xlink"
-  in
-    -- oddly, I suppose do to namespacing, the CSS class on <svg> not
-    -- picked up wrapping it seems to work though
-    H.i
-      [ P.class_ $ H.ClassName $ "sd-icon sd-icon--" <> s ]
-      [ svgElem (H.ElemName "svg")
-        []
-        [ svgElem (H.ElemName "use")
-          [ xlinkAttr (H.AttrName "xlink:href") $ "#sd-icon--" <> s ]
-          []
-        ]
-      ]
-
-${html.join("\n\n")}
-`;
-
-    fs.writeFileSync(path.join("src", "SlamData", "Render", "Icon.purs"), pursData);
-  })
-);
-
-
-gulp.task("icons", [ "svg-icons", "icon-purs" ]);
+gulp.task("inject-icons", injectIconsIntoHTML);
+gulp.task("icon-purs", createIconPureScript);
+gulp.task("icons", [ "inject-icons", "icon-purs" ]);
 gulp.task("full", [ "clean", "add-headers", "trim-whitespace", "less", "icons", "make-bundle" ]);
 gulp.task("default", [ "less", "icons", "make-bundle" ]);
